[EL] Authentication fixes

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/ShoppingListServiceImpl.java
@@ -8,7 +8,6 @@ import at.ac.tuwien.sepr.groupphase.backend.endpoint.mapper.ItemMapper;
 import at.ac.tuwien.sepr.groupphase.backend.endpoint.mapper.LabelMapper;
 import at.ac.tuwien.sepr.groupphase.backend.endpoint.mapper.ShoppingListMapper;
 import at.ac.tuwien.sepr.groupphase.backend.entity.ApplicationUser;
-import at.ac.tuwien.sepr.groupphase.backend.entity.DigitalStorage;
 import at.ac.tuwien.sepr.groupphase.backend.entity.Ingredient;
 import at.ac.tuwien.sepr.groupphase.backend.entity.Item;
 import at.ac.tuwien.sepr.groupphase.backend.entity.ItemLabel;
@@ -16,6 +15,7 @@ import at.ac.tuwien.sepr.groupphase.backend.entity.ShoppingItem;
 import at.ac.tuwien.sepr.groupphase.backend.entity.ShoppingList;
 import at.ac.tuwien.sepr.groupphase.backend.exception.AuthenticationException;
 import at.ac.tuwien.sepr.groupphase.backend.exception.ConflictException;
+import at.ac.tuwien.sepr.groupphase.backend.exception.ValidationException;
 import at.ac.tuwien.sepr.groupphase.backend.repository.ItemRepository;
 import at.ac.tuwien.sepr.groupphase.backend.repository.ShoppingListRepository;
 import at.ac.tuwien.sepr.groupphase.backend.repository.ShoppingRepository;
@@ -23,15 +23,16 @@ import at.ac.tuwien.sepr.groupphase.backend.service.DigitalStorageService;
 import at.ac.tuwien.sepr.groupphase.backend.service.ItemService;
 import at.ac.tuwien.sepr.groupphase.backend.service.LabelService;
 import at.ac.tuwien.sepr.groupphase.backend.service.ShoppingListService;
+import at.ac.tuwien.sepr.groupphase.backend.service.impl.authenticator.Authorization;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.stereotype.Service;
 
 import java.lang.invoke.MethodHandles;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -48,10 +49,13 @@ public class ShoppingListServiceImpl implements ShoppingListService {
     private final ItemService itemService;
     private CustomUserDetailService customUserDetailService;
 
+    private final Authorization authorization;
+    private final SharedFlatService sharedFlatService;
+
     public ShoppingListServiceImpl(ShoppingRepository shoppingRepository, ShoppingListRepository shoppingListRepository,
                                    ShoppingListMapper shoppingListMapper, LabelService labelService, ItemMapper itemMapper,
                                    IngredientMapper ingredientMapper, ItemRepository itemRepository, DigitalStorageService digitalStorageService,
-                                   ItemService itemService, LabelMapper labelMapper, CustomUserDetailService customUserDetailService) {
+                                   ItemService itemService, LabelMapper labelMapper, CustomUserDetailService customUserDetailService, Authorization authorization, SharedFlatService sharedFlatService) {
         this.shoppingRepository = shoppingRepository;
         this.labelService = labelService;
         this.itemMapper = itemMapper;
@@ -62,11 +66,17 @@ public class ShoppingListServiceImpl implements ShoppingListService {
         this.digitalStorageService = digitalStorageService;
         this.itemService = itemService;
         this.customUserDetailService = customUserDetailService;
+        this.authorization = authorization;
+        this.sharedFlatService = sharedFlatService;
     }
 
     @Override
-    public ShoppingItem create(ShoppingItemDto itemDto) {
-        LOGGER.trace("create({})", itemDto);
+    public ShoppingItem create(ShoppingItemDto itemDto, String jwt) throws AuthenticationException {
+        LOGGER.trace("create({},{})", itemDto, jwt);
+        ApplicationUser applicationUser = customUserDetailService.getUser(jwt);
+        if (applicationUser == null) {
+            throw new AuthenticationException("Authentication failed", List.of("User does not exist"));
+        }
         List<ItemLabel> labels = findLabelsAndCreateMissing(itemDto.labels());
 
         ShoppingItem createdItem = shoppingRepository.save(itemMapper.dtoToShopping(itemDto, labels, shoppingListMapper.dtoToEntity(itemDto.shoppingList())));
@@ -75,47 +85,105 @@ public class ShoppingListServiceImpl implements ShoppingListService {
     }
 
     @Override
-    public Optional<ShoppingItem> getById(Long itemId) {
-        LOGGER.trace("getById({})", itemId);
+    public Optional<ShoppingItem> getById(Long itemId,String jwt) throws AuthenticationException {
+        LOGGER.trace("getById({},{})", itemId, jwt);
+        ApplicationUser applicationUser = customUserDetailService.getUser(jwt);
+        if (applicationUser == null) {
+            throw new AuthenticationException("Authentication failed", List.of("User does not exist"));
+        }
         if (itemId == null) {
             return Optional.empty();
+        }
+        Optional<ShoppingItem> itemOptional = shoppingRepository.findById(itemId);
+        if (itemOptional.isPresent()) {
+            ShoppingItem item = itemOptional.get();
+            if (item.getShoppingList().getSharedFlat().equals(applicationUser.getSharedFlat())) {
+                throw new AuthenticationException("Autentication error", List.of("User has no access to this shopping items"));
+            }
         }
 
         return shoppingRepository.findById(itemId);
     }
 
     @Override
-    public Optional<ShoppingList> getShoppingListById(Long shopListId) {
-        LOGGER.trace("getShoppingListById({})", shopListId);
+    public Optional<ShoppingList> getShoppingListById(Long shopListId, String jwt) throws AuthenticationException {
+        LOGGER.trace("getShoppingListById({},{})", shopListId, jwt);
+        ApplicationUser applicationUser = customUserDetailService.getUser(jwt);
+        if (applicationUser == null) {
+            throw new AuthenticationException("Authentication failed", List.of("User does not exist"));
+        }
         if (shopListId == null) {
             return Optional.empty();
         }
-
+        Optional<ShoppingList> listOptional = shoppingListRepository.getByShopListId(shopListId);
+        if (listOptional.isPresent()) {
+            ShoppingList list = listOptional.get();
+            if (list.getSharedFlat().equals(applicationUser.getSharedFlat())) {
+                throw new AuthenticationException("Autentication error", List.of("User has no access to this shopping list"));
+            }
+        }
         return shoppingListRepository.getByShopListId(shopListId);
     }
 
     @Override
-    public List<ShoppingItem> getItemsById(Long listId, ShoppingItemSearchDto itemSearchDto) {
-        LOGGER.trace("getItemsById({}, {})", listId, itemSearchDto);
+    public List<ShoppingItem> getItemsById(Long listId, ShoppingItemSearchDto itemSearchDto, String jwt) throws AuthenticationException {
+        LOGGER.trace("getItemsById({},{},{})", listId, itemSearchDto, jwt);
+        ApplicationUser applicationUser = customUserDetailService.getUser(jwt);
+        if (applicationUser == null) {
+            throw new AuthenticationException("Authentication failed", List.of("User does not exist"));
+        }
         List<ShoppingItem> shoppingItems = shoppingRepository.searchItems(listId,
             (itemSearchDto.productName() != null) ? itemSearchDto.productName() : null,
             (itemSearchDto.label() != null) ? itemSearchDto.label() : null);
+        for (ShoppingItem item:shoppingItems) {
+            if (!item.getShoppingList().getSharedFlat().equals(applicationUser.getSharedFlat())){
+                throw new AuthenticationException("Authentication error", List.of("This user has no access to these items"));
+            }
+        }
         return shoppingRepository.saveAll(shoppingItems);
     }
 
     @Override
-    public ShoppingList createList(String listName) {
+    public ShoppingList createList(String listName, String jwt) throws ValidationException, AuthenticationException {
+        LOGGER.trace("createList({},{})", listName, jwt);
+        ApplicationUser applicationUser = customUserDetailService.getUser(jwt);
+        if (applicationUser == null) {
+            throw new AuthenticationException("Authentication failed", List.of("User does not exist"));
+        }
+
+        if (Objects.equals(listName, "Default")) {
+            throw new ValidationException("List names can not be Default!", null);
+        }
         ShoppingList shoppingList = new ShoppingList();
         shoppingList.setName(listName);
+        List<Long> allowedUser = sharedFlatService.findById(
+                shoppingList.getSharedFlat().getId(),
+                jwt
+            ).getUsers().stream()
+            .map(ApplicationUser::getId)
+            .toList();
+        authorization.authenticateUser(
+            jwt,
+            allowedUser,
+            "The given shopping list does not belong to the user's shared flat!"
+        );
         return shoppingListRepository.save(shoppingList);
     }
 
     @Override
-    public ShoppingItem deleteItem(Long itemId) {
+    public ShoppingItem deleteItem(Long itemId, String jwt) throws AuthenticationException {
+        LOGGER.trace("deleteItem({},{})", itemId, jwt);
+        ApplicationUser applicationUser = customUserDetailService.getUser(jwt);
+        if (applicationUser == null) {
+            throw new AuthenticationException("Authentication failed", List.of("User does not exist"));
+        }
         Optional<ShoppingItem> toDeleteOptional = shoppingRepository.findById(itemId);
 
         if (toDeleteOptional.isPresent()) {
             ShoppingItem toDelete = toDeleteOptional.get();
+            if (!toDelete.getShoppingList().getSharedFlat().equals(applicationUser.getSharedFlat())) {
+                throw new AuthenticationException("Authentication wrong", List.of("User can not delete this item"));
+            }
             shoppingRepository.deleteById(itemId);
             return toDelete;
         } else {
@@ -124,14 +192,21 @@ public class ShoppingListServiceImpl implements ShoppingListService {
     }
 
     @Override
-    public ShoppingList deleteList(Long shopId) {
-        if (shopId == 1) {
-            throw new BadCredentialsException("The main list can not be deleted!");
+    public ShoppingList deleteList(Long shopId, String jwt) throws ValidationException, AuthenticationException {
+        LOGGER.trace("deleteList({},{})", shopId, jwt);
+        ApplicationUser applicationUser = customUserDetailService.getUser(jwt);
+        if (applicationUser == null) {
+            throw new AuthenticationException("Authentication failed", List.of("User does not exist"));
         }
         Optional<ShoppingList> toDeleteOptional = shoppingListRepository.findById(shopId);
-
         if (toDeleteOptional.isPresent()) {
             ShoppingList toDelete = toDeleteOptional.get();
+            if (!toDelete.getSharedFlat().equals(applicationUser.getSharedFlat())) {
+                throw new AuthenticationException("Authentication wrong", List.of("User can not delete this list"));
+            }
+            if (toDelete.getName().equals("Default")) {
+                throw new ValidationException("Default list can not be deleted!", null);
+            }
             List<ShoppingItem> items = shoppingRepository.findByShoppingListId(shopId);
             shoppingRepository.deleteAll(items);
             shoppingListRepository.deleteById(shopId);
@@ -152,15 +227,19 @@ public class ShoppingListServiceImpl implements ShoppingListService {
     }
 
     @Override
-    public List<Item> transferToServer(List<ShoppingItemDto> items) {
-        LOGGER.trace("transferToServer({})", items);
+    public List<Item> transferToServer(List<ShoppingItemDto> items, String jwt) throws AuthenticationException {
+        LOGGER.trace("transferToServer({},{})", items, jwt);
+        ApplicationUser applicationUser = customUserDetailService.getUser(jwt);
+        if (applicationUser == null) {
+            throw new AuthenticationException("Authentication failed", List.of("User does not exist"));
+        }
         List<Item> itemsList = new ArrayList<>();
         for (ShoppingItemDto itemDto : items) {
             Item item;
             if (itemDto.alwaysInStock() != null && itemDto.alwaysInStock()) {
-                item = shoppingListMapper.shoppingItemDtoToAis(itemDto, ingredientMapper.dtoListToEntityList(itemDto.ingredients()), new DigitalStorage());
+                item = shoppingListMapper.shoppingItemDtoToAis(itemDto, ingredientMapper.dtoListToEntityList(itemDto.ingredients()), applicationUser.getSharedFlat().getDigitalStorage());
             } else {
-                item = shoppingListMapper.shoppingItemDtoToItem(itemDto, ingredientMapper.dtoListToEntityList(itemDto.ingredients()), new DigitalStorage());
+                item = shoppingListMapper.shoppingItemDtoToItem(itemDto, ingredientMapper.dtoListToEntityList(itemDto.ingredients()), applicationUser.getSharedFlat().getDigitalStorage());
             }
             itemRepository.save(item);
             shoppingRepository.deleteById(itemDto.itemId());
